requirements.txt
# Idea Inc - Python Dependencies
# Generated from pyproject.toml for compatibility

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Authentication
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
httpx>=0.26.0
authlib>=1.3.0

# Databases
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0
motor>=3.3.0
pymongo>=4.6.0
redis>=5.0.0

# Event Bus
aiokafka>=0.10.0

# Vector DB
chromadb>=0.4.22

# AI/LLM
openai>=1.10.0
tiktoken>=0.5.0

# Observability
structlog>=24.1.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0
opentelemetry-exporter-otlp>=1.22.0
prometheus-client>=0.19.0

# GraphQL
strawberry-graphql[fastapi]>=0.217.0

# Utilities
python-dotenv>=1.0.0
email-validator>=2.1.0
certifi>=2024.0.0
orjson>=3.9.0
numpy>=1.26.0
networkx>=3.2.0

# Development (optional)
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
ruff>=0.1.0
mypy>=1.8.0

# gRPC (optional)
grpcio>=1.60.0
grpcio-tools>=1.60.0

